Use latest mynt (0.3.1)
The pyladies website was explicitly using mynt 0.2.3,
which depends on a library called misaka. However misaka
had been updated and the latest version no longer worked
with mynt 0.2.3.

This commit updates the pyladies site to work with mynt
0.3.1, which doesn't use misaka at all. mynt made a
number of backwards incompatible changes (e.g. the
local variable "post" is now called "item"), some
of which can be seen on the mynt github page:
https://github.com/Anomareh/mynt/blob/master/NEWS.md
It also didn't recognize files with the .mdown
extension so I changed those to .md.

Also fixed some blog posts which meant to set tags
but used "tag" instead of "tags".